% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rnaiveols-package.R
\name{cpp4rols-package}
\alias{cpp4rols-package}
\title{Obtain the OLS Estimator With a Naive Computation}
\description{
How to use 'cpp4r' and obtain the OLS estimator in the most naive way. There is a
 good explanation about why R's lm() function uses a QR decomposition instead of directly
 computing (X'X)-1(X'Y) as I did here. See
 \url{http://madrury.github.io/jekyll/update/statistics/2016/07/20/lm-in-R.html}.
}

